cleaning up, removing some debug logs
